% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{directQuery}
\alias{directQuery}
\title{Direct query to database}
\usage{
directQuery(database, query)
}
\arguments{
\item{database}{name of the database as a string}

\item{query}{the query as a string}
}
\value{
the requested data
}
\description{
Direct query to database
}
\examples{
directQuery(database = 'database_name', query = 'SELECT * FROM table')
}
