% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droptable.R
\name{sql_droptable}
\alias{sql_droptable}
\title{Deletes data from tables with SQL queries}
\usage{
sql_droptable(table_name)
}
\arguments{
\item{table}{the table to delete from}

\item{where_condition}{a string of conditions used to specify, need to use escape key for quotes}
}
\description{
Deletes data from tables with SQL queries
}
\examples{
sql_delete("people", c("name = \\"Maria\\" OR age = 28 AND gender = \\"female\\"")
sql_delete("car", c("make = \\"Toyota\\" AND model = \\"Camry\\""))
}

