% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update.R
\name{sql_update}
\alias{sql_update}
\title{Updates the table with parameter}
\usage{
sql_update(table_name, columns = -1, values = -1, where_column = -1,
  where_value = -1)
}
\arguments{
\item{columns}{a list of columns to update}

\item{values}{a list of values to be updated that must be the same size as the columns}

\item{table}{connection where the function is  acting on}
}
\description{
Updates the table with parameter
}
\examples{
sql_delete("people", c("name = \\"Maria\\" OR age = 28 AND gender = \\"female\\"")
sql_delete("car", c("make = \\"Toyota\\" AND model = \\"Camry\\""))
}

