% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sqliter-package}
\alias{sqliter-package}
\title{Functions to wrap SQLite calls}
\description{
sqliter helps users, mainly data munging practioneers, to organize
their sql calls in a clean structure. It simplifies the process of
extracting and transforming data into useful formats.
}

